refactor: modularize app shell, editor helpers, and styles
