+++ src/extensions/keyboard/boldShortcut.js
@@ -1,0 +1,10 @@
+import { Extension } from '@tiptap/core'
+
+export const BoldShortcut = Extension.create({
+  name: 'boldShortcut',
+  addKeyboardShortcuts() {
+    return {
+      'Mod-b': () => this.editor.chain().focus().toggleBold().run(),
+    }
+  },
+})
