+++ src/extensions/keyboard/index.js
@@ -1,0 +1,8 @@
+export { LinkShortcut } from './linkShortcut'
+export { BoldShortcut } from './boldShortcut'
+export { ArrowMoveToLineEnd } from './arrowMoveToLineEnd'
+export { ListIndentShortcut } from './listIndentShortcut'
+export { ListSelectShortcut } from './listSelectShortcut'
+export { ListExitOnEmpty } from './listExitOnEmpty'
+export { ListEnterOutdent } from './listEnterOutdent'
+export { ListBackspaceOutdent } from './listBackspaceOutdent'
