+++ src/styles/index.css
@@ -1,0 +1,6 @@
+@import './base.css';
+@import './layout.css';
+@import './toolbar.css';
+@import './menus-modals.css';
+@import './editor.css';
+@import './responsive.css';
