+++ src/styles/layout.css
@@ -1,0 +1,247 @@
+.app {
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  color: #0f172a;
+}
+
+.app-auth {
+  max-width: 720px;
+  margin: 0 auto;
+  padding: 3rem 1.5rem 4rem;
+  gap: 2rem;
+}
+
+.topbar {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  gap: 1rem;
+  padding: 1.5rem 2.5rem 0.5rem;
+}
+
+.topbar-left {
+  display: flex;
+  align-items: flex-start;
+  gap: 2rem;
+  flex-wrap: wrap;
+}
+
+.topbar-actions {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+}
+
+.settings-button.active {
+  border-color: #2563eb;
+  color: #2563eb;
+}
+
+.brand h1 {
+  margin: 0 0 0.25rem;
+}
+
+.notebook-switcher {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  flex-wrap: wrap;
+}
+
+.notebook-switcher select {
+  border: 1px solid #e2e8f0;
+  border-radius: 8px;
+  padding: 0.35rem 0.5rem;
+  background: #ffffff;
+  font-size: 0.9rem;
+}
+
+.section-tabs {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  padding: 0 2.5rem 0.5rem;
+  overflow-x: auto;
+}
+
+.section-tab {
+  border: 1px solid #e2e8f0;
+  border-bottom: none;
+  border-radius: 12px 12px 0 0;
+  padding: 0.4rem 0.8rem;
+  background: #eef2ff;
+  font-size: 0.85rem;
+  display: inline-flex;
+  align-items: center;
+  gap: 0.5rem;
+  cursor: pointer;
+  user-select: none;
+}
+
+.section-tab.active {
+  border-color: #2563eb;
+  font-weight: 600;
+}
+
+.tab-delete {
+  border: none;
+  background: transparent;
+  color: #64748b;
+  font-size: 0.9rem;
+  padding: 0;
+  cursor: pointer;
+}
+
+.section-add {
+  border: 1px dashed #cbd5f5;
+  background: transparent;
+  color: #2563eb;
+  border-radius: 10px;
+  padding: 0.35rem 0.6rem;
+  cursor: pointer;
+}
+
+.welcome {
+  padding: 1.5rem 2.5rem 2.5rem;
+}
+
+.workspace {
+  flex: 1;
+  display: grid;
+  grid-template-columns: 1fr 220px;
+  gap: 1.25rem;
+  padding: 0 2.5rem 2.5rem;
+  align-items: start;
+}
+
+.workspace > * {
+  min-width: 0;
+}
+
+.workspace.settings-mode {
+  grid-template-columns: 1fr;
+}
+
+.sidebar {
+  border-left: 1px solid #e2e8f0;
+  padding: 1.5rem 0 1.5rem 1rem;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  height: fit-content;
+}
+
+.sidebar-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 0.5rem;
+}
+
+.sidebar-header h2 {
+  margin: 0;
+  font-size: 0.95rem;
+  text-transform: uppercase;
+  letter-spacing: 0.08em;
+  color: #64748b;
+}
+
+.sidebar-list {
+  display: flex;
+  flex-direction: column;
+  gap: 0.35rem;
+}
+
+.sidebar-row {
+  display: flex;
+  align-items: center;
+  gap: 0.45rem;
+}
+
+.sidebar-item {
+  border-left: 2px solid transparent;
+  border-radius: 6px;
+  padding: 0.45rem 0.6rem;
+  text-align: left;
+  background: transparent;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  color: #334155;
+}
+
+.sidebar-item.active {
+  border-color: #2563eb;
+  color: #0f172a;
+  background: #eff6ff;
+}
+
+.sidebar-item.drag-over {
+  border-color: #2563eb;
+  background: #eef2ff;
+}
+
+.sidebar-title {
+  font-weight: 500;
+  font-size: 0.9rem;
+  min-width: 0;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.tracker-page-badge {
+  font-size: 0.6rem;
+  line-height: 1;
+  letter-spacing: 0.06em;
+  color: #1d4ed8;
+  background: #dbeafe;
+  border: 1px solid #93c5fd;
+  border-radius: 999px;
+  padding: 0.18rem 0.35rem;
+  white-space: nowrap;
+}
+
+.settings-hub {
+  background: #ffffff;
+  border: 1px solid #e2e8f0;
+  border-radius: 14px;
+  padding: 1.5rem;
+  box-shadow: 0 12px 30px rgba(15, 23, 42, 0.08);
+}
+
+.settings-header {
+  display: flex;
+  align-items: flex-start;
+  justify-content: space-between;
+  gap: 1rem;
+  margin-bottom: 1.5rem;
+}
+
+.settings-grid {
+  display: grid;
+  gap: 1rem;
+}
+
+.settings-card {
+  border: 1px solid #e2e8f0;
+  border-radius: 12px;
+  padding: 1.25rem;
+  background: #f8fafc;
+  display: flex;
+  align-items: flex-start;
+  justify-content: space-between;
+  gap: 1rem;
+  flex-wrap: wrap;
+}
+
+.settings-card h3 {
+  margin: 0 0 0.35rem;
+  font-size: 1rem;
+}
+
+.settings-loading {
+  color: #64748b;
+}
